xcopy-populator: refactor vmkfstools-wrapper to be job based
MOTIVATION
Long running copy execution times can timeout the vsphere->esxi
communication channel, and fail the populator execution (while the copy
may still be in progress).
Also, there is no progress reporting by vmkfstool execution, due to it being
syncronic with no way to track it other than the std out of the command.

MODIFICAITON
Instead of a call to the wrapper that would wait till vmkfstools return,
the wrapper now allocates a temp directory with a uuid, and then calls
vmkfstools and putting it in the background, with its own parent id, so
it can fork.
Also the invocation captures the Pid, stdout, stderr, and exitcode of the task
and directs them into files under that folder: Example:

root@ecoesxi07:~] ls -la /tmp/vmkfstools-wrapper-ca0d8704-8f49-45ad-b071-09a7124428cd/
total 20
drwxr-x---    1 root     root           512 May  6 07:51 .
drwx------    1 root     root           512 May  6 14:01 ..
-rw-r--r--    1 root     root             0 May  6 07:49 err
-rw-r--r--    1 root     root             1 May  6 07:51 exitcode
-rw-r--r--    1 root     root          1874 May  6 07:51 out
-rw-r--r--    1 root     root             8 May  6 07:49 pid

Every task now has its state captures on disk and the wrapper returns
the call to clone with the task ID.
The populator calling this wrapper now reserve this task ID and can
track it using a new call to get the task ID using:

   /bin/vmkftools-wrapper -i ca0d8704-8f49-45ad-b071-09a7124428cd

This will return a json representation of with all the needed info. The
populator code will stop monitoring when the task has an exit code.
Meanwhile, it uses the reported stdout, parse it, and report the
progress with it.

CHANGE
Vib version 0.0.77 must be installed, and it now has a breaking change. It
returns immediately with the task ID in the response, encapsulated in a
json format.
The wrapper clone verb didn't change.
The wrapper has a new verb, task-get, to fetch the task status.
To match this change the populator remote_esxcli.go code is now
monitoring the task with the new flag added to the wrapper.
The vmkfstools-wrapper is re-written in python and properly logs its
stuff to /var/log/vmkfstools-wrapper.log

Signed-off-by: Roy Golan <rgolan@redhat.com>